+++ src/aws-api-mcp-server/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for `--endpoint-url` flag for localhost endpoints (#1452)
+- Change max retries to 3 when interpreting CLI command (#1485)
 
 ## [1.0.1] - 2025-10-06